Update the database folder to be organized according to the MVC pattern
