+++ src/routes/users/profile/profile.js
@@ -1,8 +1,8 @@
 const express = require('express');
 
 const router = express.Router();
-const reviewFunctions = require('../../../database/controllers/review');
-const userFunctions = require('../../../database/controllers/user');
+const reviewFunctions = require('../../../models/db/review');
+const userFunctions = require('../../../models/db/user');
 
 
 router.get('/', (request, response) => {
